requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
config==0.5.1
decorator==5.2.1
dm-tree==0.1.8
exceptiongroup==1.3.0
faiss-cpu==1.8.0.post1
fastapi==0.115.12
filelock==3.16.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
h5py==3.11.0
huggingface-hub==0.31.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.13.1
libclang==18.1.1
llm_providers==0.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.39.1
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
propcache==0.2.0
protobuf==4.25.7
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.24.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-probability==0.21.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.4.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
Werkzeug==3.0.6
wrapt==1.17.2
yarl==1.15.2
zipp==3.20.2
